+++ build_venv.sh
@@ -1,5 +1,0 @@
-sudo apt-get install -y virtualenv
-sudo apt-get install -y python3 python3-pip python3-dev
-
-mkdir pyenv
-virtualenv pyenv --python=python3