+++ manage.py
@@ -8,7 +8,6 @@ from flask_alembic import alembic_click
 from flask_migrate.cli import db as db_cli
 
 from app import create_app
-from app.commom import utils
 from tests.v1.api import build_api_test_suite
 
 
@@ -24,17 +23,8 @@ def run():
     app.run(host="0.0.0.0")
 
 
-@click.command()
-def test():
-    suite = build_api_test_suite()
-    with open('UnittestTextReport.txt', 'w') as f:
-        runner = unittest.TextTestRunner(stream=f, verbosity=2)
-        runner.run(suite)
-
-
 cli.add_command(run)
 cli.add_command(db_cli)
-cli.add_command(test)
 
 if __name__ == "__main__":
     cli()